Add `ignoreeof` flag to socat tcp listener (#22)
After a maddening few hours of using `socat` as a proxy to the docker socket and wondering why `docker run` would inconsistently fail to return any output I found it was due to the unfortunate behaviour of `socat` and the `docker` cli. I believe that the docker client is leaving the socket half-open which triggers socat to timeout (after 0.5s by default) if there is no data to read from the opposite direction. Increasing the timeout (`socat -t10`) worked but ignoring the EOF entirely seems more robust.